feat: implement consolidated bill section and refactor dashboard components
