+++ src/components/dashboard/index.ts
@@ -1,3 +1,5 @@
+export * from "@/components/dashboard/bill-breakdown";
+export * from "@/components/dashboard/consolidated-bill-section";
 export * from "@/components/dashboard/dashboard-page";
 export * from "@/components/dashboard/dashboard-wrapper";
 export * from "@/components/dashboard/stat-card";
